# videodna-studio/packages/api-client/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "VideoDNA Studio API",
    "description": "API do VideoDNA Studio — análise, desmontagem semântica, edição estrutural e reconstrução de vídeos com múltiplos modelos de IA. Erros seguem o formato `{error: {code, message, details, requestId}}` com códigos normalizados.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/entities/{entity_id}/suggestions": {
      "get": {
        "tags": [
          "suggestions"
        ],
        "summary": "Sugestões contextuais para um elemento (cacheadas)",
        "operationId": "get_suggestions_entities__entity_id__suggestions_get",
        "parameters": [
          {
            "name": "entity_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Entity Id"
            }
          },
          {
            "name": "category",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Category"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 20,
              "minimum": 0,
              "default": 0,
              "title": "Page"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuggestionList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/entities/{entity_id}/suggestions/more": {
      "post": {
        "tags": [
          "suggestions"
        ],
        "summary": "Gerar mais opções (próxima página)",
        "operationId": "more_suggestions_entities__entity_id__suggestions_more_post",
        "parameters": [
          {
            "name": "entity_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Entity Id"
            }
          },
          {
            "name": "category",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Category"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuggestionList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/suggestions/{suggestion_id}/apply": {
      "post": {
        "tags": [
          "suggestions"
        ],
        "summary": "Aplicar uma sugestão (cria a operação de edição)",
        "operationId": "apply_suggestion_suggestions__suggestion_id__apply_post",
        "parameters": [
          {
            "name": "suggestion_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Suggestion Id"
            }
          },
          {
            "name": "instruction",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 2000
                },
                {
                  "type": "null"
                }
              ],
              "title": "Instruction"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EditResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/suggestions/{suggestion_id}/preview": {
      "post": {
        "tags": [
          "suggestions"
        ],
        "summary": "Preview de frame (a sugestão aplicada a um keyframe real)",
        "operationId": "preview_suggestion_suggestions__suggestion_id__preview_post",
        "parameters": [
          {
            "name": "suggestion_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Suggestion Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/PreviewRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Req"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChangeKind": {
        "type": "string",
        "enum": [
          "METADATA",
          "ATTRIBUTE_COLOR",
          "LOCAL_APPEARANCE",
          "OBJECT_REPLACE",
          "OBJECT_REMOVE",
          "OBJECT_ADD",
          "ENVIRONMENT_PART",
          "ENVIRONMENT_FULL",
          "CHARACTER_APPEARANCE",
          "CHARACTER_REPLACE",
          "CHARACTER_REMOVE",
          "INSTRUCTION"
        ],
        "title": "ChangeKind"
      },
...
      "Dependency": {
        "properties": {
          "type": {
            "$ref": "#/components/schemas/DependencyType"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "entityId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Entityid"
          },
          "actionId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actionid"
          },
          "shotIds": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Shotids"
          },
          "futureFeature": {
            "type": "boolean",
            "title": "Futurefeature",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "type",
          "description",
          "entityId",
          "actionId",
          "shotIds",
          "futureFeature"
        ],
        "title": "Dependency"
      },
      "DependencyType": {
        "type": "string",
        "enum": [
          "HAND_INTERACTION",
          "PHYSICS_MOTION",
          "DESTRUCTION_FX",
          "CONTACT_SURFACE",
          "GAZE_TARGET",
          "OCCLUSION",
          "CONTINUITY",
          "DERIVED_ENTITY",
          "LIGHTING",
          "STORY_ROLE",
          "AUDIO_SFX",
          "CHILD_ELEMENTS"
        ],
        "title": "DependencyType"
      },
...
      "EditHistoryOut": {
        "properties": {
          "edits": {
            "items": {
              "$ref": "#/components/schemas/EditOut"
            },
            "type": "array",
            "title": "Edits"
          },
          "canUndo": {
            "type": "boolean",
            "title": "Canundo"
          },
          "canRedo": {
            "type": "boolean",
            "title": "Canredo"
          }
        },
        "type": "object",
        "required": [
          "edits",
          "canUndo",
          "canRedo"
        ],
        "title": "EditHistoryOut"
      },
      "EditOpType": {
        "type": "string",
        "enum": [
          "SET_ATTRIBUTE",
          "CHANGE_APPEARANCE",
          "REPLACE",
          "REMOVE",
          "KEEP",
          "APPLY_PRESET",
          "ADD_ENTITY",
          "CORRECT",
          "INSTRUCTION"
        ],
        "title": "EditOpType"
      },
      "EditOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "sequence": {
            "type": "integer",
            "title": "Sequence"
          },
          "entityKey": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Entitykey"
          },
          "op": {
            "$ref": "#/components/schemas/EditOpType"
          },
          "property": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Property"
          },
          "previousValue": {
            "title": "Previousvalue"
          },
          "newValue": {
            "title": "Newvalue"
          },
          "instruction": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instruction"
          },
          "source": {
            "$ref": "#/components/schemas/EditSource"
          },
          "state": {
            "$ref": "#/components/schemas/EditState"
          },
          "impact": {
            "additionalProperties": true,
            "type": "object",
            "title": "Impact"
          },
          "suggestionId": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Suggestionid"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat"
          }
        },
        "type": "object",
        "required": [
          "id",
          "sequence",
          "entityKey",
          "op",
          "property",
          "previousValue",
          "newValue",
          "instruction",
          "source",
          "state",
          "impact",
          "suggestionId",
          "createdAt"
        ],
        "title": "EditOut"
      },
      "EditResult": {
        "properties": {
          "edit": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EditOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "impact": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ImpactReport"
              },
              {
                "type": "null"
              }
            ]
          },
          "history": {
            "$ref": "#/components/schemas/EditHistoryOut"
          }
        },
        "type": "object",
        "required": [
          "edit",
          "impact",
          "history"
        ],
        "title": "EditResult"
      },
      "EditSource": {
        "type": "string",
        "enum": [
          "manual",
          "suggestion",
          "ai"
        ],
        "title": "EditSource"
      },
      "EditState": {
        "type": "string",
        "enum": [
          "ACTIVE",
          "UNDONE",
          "DISCARDED"
        ],
        "title": "EditState"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ImpactLevel": {
        "type": "string",
        "enum": [
          "NONE",
          "LOW",
          "MEDIUM",
          "HIGH"
        ],
        "title": "ImpactLevel"
      },
      "ImpactReport": {
        "properties": {
          "operationId": {
            "type": "string",
            "title": "Operationid"
          },
          "entityId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Entityid"
          },
          "changeKind": {
            "$ref": "#/components/schemas/ChangeKind"
          },
          "level": {
            "$ref": "#/components/schemas/ImpactLevel"
          },
          "score": {
            "type": "number",
            "title": "Score"
          },
          "reasons": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Reasons"
          },
          "affectedShotIds": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Affectedshotids"
          },
          "affectedEntityIds": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Affectedentityids"
          },
          "dependencies": {
            "items": {
              "$ref": "#/components/schemas/Dependency"
            },
            "type": "array",
            "title": "Dependencies"
          },
          "warnings": {
            "items": {
              "$ref": "#/components/schemas/ImpactWarning"
            },
            "type": "array",
            "title": "Warnings"
          }
        },
        "type": "object",
        "required": [
          "operationId",
          "entityId",
          "changeKind",
          "level",
          "score",
          "reasons",
          "affectedShotIds",
          "affectedEntityIds",
          "dependencies",
          "warnings"
        ],
        "title": "ImpactReport"
      },
      "ImpactWarning": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "blocking": {
            "type": "boolean",
            "title": "Blocking",
            "default": false
          },
          "entityId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Entityid"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message",
          "blocking",
          "entityId"
        ],
        "title": "ImpactWarning"
      },
...
      "JobKind": {
        "type": "string",
        "enum": [
          "INGEST",
          "ANALYSIS",
          "PREVIEW",
          "GENERATION"
        ],
        "title": "JobKind"
      },
      "JobOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "projectId": {
            "type": "string",
            "format": "uuid",
            "title": "Projectid"
          },
          "kind": {
            "$ref": "#/components/schemas/JobKind"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus"
          },
          "stage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stage"
          },
          "progress": {
            "type": "number",
            "title": "Progress",
            "default": 0.0
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "errorCode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Errorcode"
          },
          "errorMessage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Errormessage"
          },
          "result": {
            "additionalProperties": true,
            "type": "object",
            "title": "Result"
          },
          "planId": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Planid"
          },
          "attempts": {
            "type": "integer",
            "title": "Attempts",
            "default": 0
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat"
          },
          "startedAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Startedat"
          },
          "finishedAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finishedat"
          }
        },
        "type": "object",
        "required": [
          "id",
          "projectId",
          "kind",
          "status",
          "stage",
          "progress",
          "message",
          "errorCode",
          "errorMessage",
          "result",
          "planId",
          "attempts",
          "createdAt",
          "startedAt",
          "finishedAt"
        ],
        "title": "JobOut"
      },
      "JobStatus": {
        "type": "string",
        "enum": [
          "QUEUED",
          "PREPARING",
          "ANALYZING",
          "GENERATING",
          "QA",
          "REPAIRING",
          "ASSEMBLING",
          "COMPLETED",
          "FAILED",
          "CANCELLED"
        ],
        "title": "JobStatus"
      },
...
      "PreviewRequest": {
        "properties": {
          "shotKey": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Shotkey"
          }
        },
        "type": "object",
        "title": "PreviewRequest"
      },
...
      "SuggestionList": {
        "properties": {
          "entityKey": {
            "type": "string",
            "title": "Entitykey"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "cached": {
            "type": "boolean",
            "title": "Cached"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/SuggestionOut"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "entityKey",
          "category",
          "page",
          "cached",
          "items"
        ],
        "title": "SuggestionList"
      },
      "SuggestionOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "entityKey": {
            "type": "string",
            "title": "Entitykey"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "op": {
            "$ref": "#/components/schemas/EditOpType"
          },
          "property": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Property"
          },
          "value": {
            "title": "Value"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          },
          "score": {
            "type": "number",
            "title": "Score"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "rank": {
            "type": "integer",
            "title": "Rank"
          },
          "previewUrl": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Previewurl"
          },
          "provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider"
          }
        },
        "type": "object",
        "required": [
          "id",
          "entityKey",
          "category",
          "label",
          "description",
          "op",
          "property",
          "value",
          "tags",
          "score",
          "page",
          "rank",
          "previewUrl",
          "provider"
        ],
        "title": "SuggestionOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  },
...
}